Up: move `init_config?` out to dedicated helper
